split up lib-iter
