+++ lib/flux-core/src/iter/adapters/enumerate.rs
@@ -1,0 +1,5 @@
+use flux_attrs::*;
+
+#[extern_spec(core::iter)]
+#[refined_by(idx: int, inner: I)]
+struct Enumerate<I>;
