+++ lib/flux-core/src/iter/adapters/map.rs
@@ -1,0 +1,5 @@
+use flux_attrs::*;
+
+#[extern_spec(core::iter)]
+#[refined_by(inner: I)]
+struct Map<I, F>;
