+++ lib/flux-core/src/iter/adapters/mod.rs
@@ -1,0 +1,2 @@
+mod map;
+mod enumerate;
