+++ lib/flux-core/src/iter/mod.rs
@@ -1,1 +1,3 @@
+mod adapters;
 mod range;
+mod traits;
