+++ lib/flux-core/src/iter/traits/mod.rs
@@ -1,0 +1,2 @@
+mod iterator;
+mod collect;
