+++ tests/tests/lib/iterator.rs
@@ -19,19 +19,6 @@ struct Skip<I>;
 #[flux_rs::refined_by(a: A, b: B, idx: int, len: int, a_len: int)]
 struct Zip<A, B>;
 
-#[flux_rs::extern_spec(core::ops)]
-#[flux_rs::assoc(fn valid_item(self: Range<A>, item: A) -> bool { self.start <= item && item < self.end })]
-#[flux_rs::assoc(fn size(self: Range<A>) -> int { <A as Step>::size(self.start, self.end) })]
-#[flux_rs::assoc(fn done(self: Range<A>) -> bool { <A as Step>::size(self.start, self.end) <= 0})]
-#[flux_rs::assoc(fn step(self: Range<A>, other: Range<A>) -> bool { <A as Step>::can_step_forward(self.start, 1) => other.start == <A as Step>::step_forward(self.start, 1) } )]
-impl<A: Step> Iterator for Range<A> {
-    #[flux_rs::sig(
-        fn(self: &mut Range<A>[@old]) -> Option<A[old.start]>[old.start < old.end]
-            ensures self: Range<A>{r: (old.start < old.end => r.start == <A as Step>::step_forward(old.start, 1)) && r.end == old.end }
-    )]
-    fn next(&mut self) -> Option<A>;
-}
-
 #[flux_rs::extern_spec(core::iter)]
 #[flux_rs::assoc(fn size(r: Skip<I>) -> int { <I as Iterator>::size(r.inner) } )]
 #[flux_rs::assoc(fn done(r: Skip<I>) -> bool { <I as Iterator>::done(r.inner) } )]
